+++ src/main/ml-modules/root/ds/lux/document/delete.mjs
@@ -1,0 +1,9 @@
+import { handleRequest } from '../../../lib/securityLib.mjs';
+import { deleteSet } from '../../../lib/setLib.mjs';
+
+const uri = external.uri;
+
+handleRequest(function () {
+  declareUpdate();
+  return deleteSet(uri);
+});
