--- src/main/ml-modules/root/ds/lux/myCollections/set/delete.mjs
@@ -1,9 +1,0 @@
-import { handleRequest } from '../../../../lib/securityLib.mjs';
-import { deleteSet } from '../../../../lib/setLib.mjs';
-
-const uri = external.uri;
-
-handleRequest(function () {
-  declareUpdate();
-  return deleteSet(uri);
-});
